forms: improve the filesystem validation
The file system validation now returns a messaged to be used in an
alert. This message is stored within an errors object passed through
data driven forms. It is accessed by wrapping our components within a
FormSpy. This error is either an empty object of it contains an error
message under `errors.customizations.filesystem`.