Wizard: enable clicking on sidebar
The sidebar is now clickable with valid fields. This requires having a
maxStepIndex and a list of all step names. This is because of ddf.

Also, the form values may not be included in the onSubmit when clicking
through the wizard so the formApi may need to be used to get the form
state.