blueprintwizard: remove empty customizations
This commit sanitizes the form data and makes sure that the values of
the form fields are truly empty. For example, the filesystem customization
option was being initialized to an empty array. This was causing `edge-commit`
and `edge-commit` image types to break since they are not allowed
filesystem customizations.

Additionally, this commit ensures that all the fields are undefined if they have
not been set. This ensures that the blueprint is saved correctly and the below
no longer happens:

```
[customizations]
[customizations.kernel]
append = ""
[customizations.timezone]
[customizations.locale]
[customizations.firewall]
[customizations.firewall.services]
[customizations.services]
[customizations.fdo]
[customizations.ignition]
[customizations.ignition.firstboot]
```